simple-agentcore-agent/pyproject.toml
[project]
name = "sample-one-quickstart"
version = "0.1.0"
description = "Sample 1: Deploy your first AI agent to AgentCore Runtime using CDK"
requires-python = ">=3.10"
dependencies = [
    "aws-cdk-lib>=2.220.0",
    "constructs>=10.0.79",
    "boto3",
    "awscrt",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
